% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.orfs.R
\name{get.orfs}
\alias{get.orfs}
\title{get.orfs}
\usage{
get.orfs(
  single_input,
  output,
  build,
  filter = NULL,
  patient = NULL,
  study_id = ""
)
}
\arguments{
\item{single_input}{Path to a single VEP-annotated .vcf or .vcf.gz file}

\item{output}{Output path}

\item{build}{Reference genome - 37 or 38}

\item{filter}{Filter vcf FILTER column by a string match e.g. "PASS". Default: NULL}

\item{patient}{String for patient ID. Default: NULL - will take basename of vcf file}

\item{study_id}{String for study_id. Default: ""}
}
\value{
returns the log output silently
}
\description{
Wrapper function to process single VCF input with PrimeCUTR.
Writes three folders to the output directory: ./log/ ./netMHC_inputs/ ./orfs/
}
\examples{
input_vcf <- primeCUTR_example("vep_HCC1937.vcf")
get.orfs(input_vcf,"./output_dir/",build=38)
}
