% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.peptide.R
\name{get.peptide}
\alias{get.peptide}
\title{get.peptide}
\usage{
get.peptide(
  ensembl_transcript_id,
  nuc_change,
  build = 38,
  check_startgains = F
)
}
\arguments{
\item{ensembl_transcript_id}{The affected ensembl transcript ID}

\item{nuc_change}{HGVS nomenclature of DNA sequence variant e.g. c.4375C>T}

\item{build}{Which reference genome to use. Defaults to GRCh38.}

\item{check_startgains}{Performs checks of kozak strength, checks for overlapping WT uORF or in-frame overlap with coding sequence}
}
\value{
A list containing predicted neoORF and metadata
}
\description{
Predicts the neopeptide produced for a given mutation and transcript combination
}
\examples{
get.peptide("ENST00000539214","c.-61C>T",build = 38,check_startgains = TRUE)
}
