% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primeCUTR-package.R
\docType{package}
\name{primeCUTR-package}
\alias{primeCUTR}
\alias{primeCUTR-package}
\title{primeCUTR: Neopeptide Prediction from Somatic Mutations of Untranslated Regions (UTR)}
\description{
A package for neopeptide prediction from somatic mutation calls of cancer genomic sequencing including start-gain, stop-loss, missense and frameshift mutations. This package requires VEP annotated variant call format as input and returns neopeptides for downstream analysis, for example via MHC binding prediction with netMHC.
}
\author{
\strong{Maintainer}: Christopher CT Sng \email{christopher.sng@nhs.net} (\href{https://orcid.org/0000-0003-2850-1467}{ORCID})

}
\keyword{internal}
