% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primeCUTR_example.R
\name{primeCUTR_example}
\alias{primeCUTR_example}
\title{Get path to primeCUTR example}
\usage{
primeCUTR_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
primeCUTR comes bundled with some example VCFs in its `inst/extdata`
directory. This function make them easy to access.
}
\examples{
primeCUTR_example()
primeCUTR_example("vep_HCC1937.vcf")
}
